pyproject.toml
[project]
name = "auto-metrics"
version = "0.1.0"
description = "Automatic METRICS assessment using large language models."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "google-genai>=1.8.0",
    "jsonref>=1.1.0",
    "ollama>=0.4.8",
    "pydantic>=2.11.1",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "selenium>=4.30.0",
]

[project.optional-dependencies]
huggingface = [
    "accelerate>=1.6.0",
    "datasets>=3.5.1",
    "outlines>=0.2.3",
    "sentencepiece>=0.2.0",
    "transformers[huggingface]>=4.51.3",
]

# build with pdm
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
